pyproject.toml
[project]
name = "shiftpilot"
version = "0.1.0"
description = "Shift scheduling desktop application"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9",
    "setuptools>=82.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["."]
